// app/src/generated/CountryTypeView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CountryTypeView = { name: string, is_space_critter: boolean, space_creatures: boolean, generate_borders: boolean, is_enclave: boolean, fallen_empire: boolean, playable: boolean, leviathan: boolean, };